smyrna warnings clean up

+++ cmd/smyrna/gui/gui.c
@@ -515,6 +515,6 @@ void append_textview(GtkTextView * textv, const char *s, size_t bytes)
     /*set iterator to the end of the buffer */
     gtk_text_buffer_get_end_iter(gtkbuf, &endit);
     /* insert buf to the end */
-    gtk_text_buffer_insert(gtkbuf, &endit, s, bytes);
+    gtk_text_buffer_insert(gtkbuf, &endit, s, (gint)bytes);
 
 }
